# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Hospital Management System.",
        "title": "HMS API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/ekastn",
            "email": "eka@voixes.studio"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5021",
    "basePath": "/api",
    "paths": {
...
        "/records": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of all medical records.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Medical Records"
                ],
                "summary": "Get all medical records",
                "responses": {
                    "200": {
                        "description": "Medical records retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.MedicalRecordDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Failed to get medical records",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new medical record entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Medical Records"
                ],
                "summary": "Create a new medical record",
                "parameters": [
                    {
                        "description": "Medical record object to be created",
                        "name": "medicalRecord",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateMedicalRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Medical record created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation failed",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to create medical record",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/records/date-range": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of medical records within a specified date range.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Medical Records"
                ],
                "summary": "Get medical records by date range",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date (RFC3339 format)",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (RFC3339 format)",
                        "name": "end",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medical records by date range",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.MedicalRecordDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Both start and end dates are required or invalid format",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get medical records",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/records/patient/{patientId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of medical records for a specific patient.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Medical Records"
                ],
                "summary": "Get medical records by patient ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "patientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient medical records",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.MedicalRecordDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Patient ID is required",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get medical records",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/records/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a single medical record by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Medical Records"
                ],
                "summary": "Get medical record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medical Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medical record details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.MedicalRecordDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Record ID is required",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Medical record not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get medical record",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update details of an existing medical record.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Medical Records"
                ],
                "summary": "Update an existing medical record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medical Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Medical record object with updated fields",
                        "name": "medicalRecord",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateMedicalRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Medical record updated successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation failed",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Medical record not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update medical record",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a medical record entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Medical Records"
                ],
                "summary": "Delete a medical record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medical Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medical record deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Record ID is required",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete medical record",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CreateMedicalRecordRequest": {
            "description": "Request body for creating a new medical record",
            "type": "object",
            "required": [
                "description",
                "diagnosis",
                "doctorId",
                "patientId",
                "recordType",
                "treatment"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 10,
                    "example": "Patient presented with flu-like symptoms."
                },
                "diagnosis": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 5,
                    "example": "Influenza A"
                },
                "doctorId": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000003"
                },
                "notes": {
                    "type": "string",
                    "maxLength": 500,
                    "example": "Advised patient to stay hydrated."
                },
                "patientId": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000002"
                },
                "recordType": {
                    "type": "string",
                    "enum": [
                        "checkup",
                        "followup",
                        "procedure",
                        "emergency"
                    ],
                    "example": "checkup"
                },
                "treatment": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 5,
                    "example": "Prescribed Tamiflu and rest."
                }
            }
        },
...
        "domain.MedicalRecordDTO": {
            "description": "Medical record data transfer object",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2025-07-17T09:00:00Z"
                },
                "date": {
                    "type": "string",
                    "example": "2025-07-17T10:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Patient presented with flu-like symptoms."
                },
                "diagnosis": {
                    "type": "string",
                    "example": "Influenza A"
                },
                "doctorId": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000003"
                },
                "id": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000001"
                },
                "notes": {
                    "type": "string",
                    "example": "Advised patient to stay hydrated."
                },
                "patientId": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000002"
                },
                "recordType": {
                    "type": "string",
                    "example": "checkup"
                },
                "treatment": {
                    "type": "string",
                    "example": "Prescribed Tamiflu and rest."
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2025-07-17T09:00:00Z"
                }
            }
        },
...
        "domain.UpdateMedicalRecordRequest": {
            "description": "Request body for updating an existing medical record",
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 10,
                    "example": "Patient's symptoms have improved."
                },
                "diagnosis": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 5,
                    "example": "Resolved Influenza A"
                },
                "notes": {
                    "type": "string",
                    "maxLength": 500,
                    "example": "Patient is recovering well."
                },
                "recordType": {
                    "type": "string",
                    "enum": [
                        "checkup",
                        "followup",
                        "procedure",
                        "emergency"
                    ],
                    "example": "followup"
                },
                "treatment": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 5,
                    "example": "Continue current medication."
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Error message"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "utils.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}